feat: update to the class1 base version

+++ src/main/java/com/afs/restapi/Employee.java
@@ -1,16 +1,16 @@
 package com.afs.restapi;
 
 public class Employee {
-    private int id;
+    private Long id;
     private String name;
-    private int age;
+    private Integer age;
     private String gender;
-    private int salary;
+    private Integer salary;
 
     public Employee() {
     }
 
-    public Employee(int id, String name, int age, String gender, int salary) {
+    public Employee(Long id, String name, Integer age, String gender, Integer salary) {
         this.id = id;
         this.name = name;
         this.age = age;
@@ -18,11 +18,11 @@ public class Employee {
         this.salary = salary;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -34,11 +34,11 @@ public class Employee {
         this.name = name;
     }
 
-    public int getAge() {
+    public Integer getAge() {
         return age;
     }
 
-    public void setAge(int age) {
+    public void setAge(Integer age) {
         this.age = age;
     }
